+++ source/externals/clhep/src/RandBreitWigner.cc
@@ -278,7 +278,7 @@ void RandBreitWigner::fireArray ( const int size, double* vect,
 
 
 std::ostream & RandBreitWigner::put ( std::ostream & os ) const {
-  int pr=os.precision(20);
+  long pr=os.precision(20);
   std::vector<unsigned long> t(2);
   os << " " << name() << "\n";
   os << "Uvec" << "\n";
