+++ source/externals/clhep/src/RandExponential.cc
@@ -77,7 +77,7 @@ void RandExponential::fireArray( const int size, double* vect,
 }
 
 std::ostream & RandExponential::put ( std::ostream & os ) const {
-  int pr=os.precision(20);
+  long pr=os.precision(20);
   std::vector<unsigned long> t(2);
   os << " " << name() << "\n";
   os << "Uvec" << "\n";
