+++ source/externals/clhep/src/RandFlat.cc
@@ -155,7 +155,7 @@ void RandFlat::restoreEngineStatus( const char filename[] ) {
 } // restoreEngineStatus
 
 std::ostream & RandFlat::put ( std::ostream & os ) const {
-  int pr=os.precision(20);
+  long pr=os.precision(20);
   std::vector<unsigned long> t(2);
   os << " " << name() << "\n";
   os << "Uvec" << "\n";
@@ -203,7 +203,7 @@ std::istream & RandFlat::get ( std::istream & is ) {
 
 std::ostream & RandFlat::saveDistState ( std::ostream & os ) {
   os << distributionName() << "\n";
-  int prec = os.precision(20);
+  long prec = os.precision(20);
   os << "RANDFLAT staticRandomInt: " << staticRandomInt 
      << "    staticFirstUnusedBit: " << staticFirstUnusedBit << "\n";
   os.precision(prec);
