+++ source/externals/clhep/src/RandStudentT.cc
@@ -140,7 +140,7 @@ double RandStudentT::shoot( HepRandomEngine *anEngine, double a ) {
 }
 
 std::ostream & RandStudentT::put ( std::ostream & os ) const {
-  int pr=os.precision(20);
+  long pr=os.precision(20);
   std::vector<unsigned long> t(2);
   os << " " << name() << "\n";
   os << "Uvec" << "\n";
